moved things for adapter model pattern
